requirements.txt
python-dotenv
streamlit
groq
duckduckgo-search
packaging
phidata
sqlalchemy
uvicorn
yfinance
google-generativeai
